requirements.txt
click
Wavelink
discord.py
discord.py[voice]